Restructured Jo Wilder Logging
The logging functions are now all in a seperate file called log_events.js. It appears to work just fine, except for sometimes the porthole hovers get logged after going into the next room. I'm not exactly sure how to fix that, but I think it is a minor bug for now.